+++ be-traveltrove/src/main/java/com/traveltrove/betraveltrove/externalservices/auth0/models/Auth0UserRequestModel.java
@@ -1,26 +1,29 @@
-package com.traveltrove.betraveltrove.domainclient.auth0;
+package com.traveltrove.betraveltrove.externalservices.auth0.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Auth0UserResponseModel {
+public class Auth0UserRequestModel {
 
-    @JsonProperty("user_id")
-    private String userId;
-
-    @JsonProperty("email")
     private String email;
+
+    private String password;
 
     @JsonProperty("given_name")
     private String firstName;
 
     @JsonProperty("family_name")
     private String lastName;
+
+    private String connection = "Username-Password-Authentication";
+
+    @JsonProperty("verify_email")
+    private boolean verifyEmail = true;
 }
