+++ be-traveltrove/src/main/java/com/traveltrove/betraveltrove/externalservices/auth0/models/Auth0UserResponseModel.java
@@ -1,29 +1,26 @@
-package com.traveltrove.betraveltrove.domainclient.auth0;
+package com.traveltrove.betraveltrove.externalservices.auth0.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Auth0UserRequestModel {
+public class Auth0UserResponseModel {
 
+    @JsonProperty("user_id")
+    private String userId;
+
+    @JsonProperty("email")
     private String email;
-
-    private String password;
 
     @JsonProperty("given_name")
     private String firstName;
 
     @JsonProperty("family_name")
     private String lastName;
-
-    private String connection = "Username-Password-Authentication";
-
-    @JsonProperty("verify_email")
-    private boolean verifyEmail = true;
 }
